chore: inline util
